+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -101,5 +101,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-10-15T10:20:52.103Z")
+    created: Optional[datetime] = parse_datetime("2025-10-15T12:57:23.766Z")
     r"""The date and time the resource was created"""