+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K7KP6925H0C7S4BWYFVF7M8G"
+        "01K7KZ4WXD8T2A9HP3TFCEHFCE"
     )
     r"""The id of the resource"""